fix: added useMint
