+++ src/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useChainInfo';
 export * from './useAccount';
 export * from './useConnectWallet';
 export * from './useStakingDapps';
+export * from './useMint';
